Reset Drawing app
